--- packages/lib/src/components/ErrorHandling/index.ts
@@ -1,2 +1,0 @@
-export { ErrorBoundary } from "./ErrorBoundary";
-export { initializeErrorHandling } from "./initialize";
